backend/requirements.txt
# Web Framework & Extensions
Flask==3.1.0
Flask-Cors==5.0.1
Flask-JWT-Extended==4.7.1
Werkzeug==3.1.3

# Firebase
firebase-admin==6.2.0
google-cloud-storage==2.14.0

# HTTP & Requests
requests==2.31.0
urllib3==2.2.0

# Utils
python-dotenv==1.1.0
PyJWT==2.8.0


# Crypto (用於加密)
cryptography==41.0.4